+++ examples/coralogix-aws-shipper/variables.tf
@@ -329,6 +329,23 @@ variable "custom_metadata" {
   type        = string
 }
 
+variable "lambda_assume_role_arn" {
+  default     = null
+  description = "The ARN of the role that the lambda function will assume. "
+  type        = string
+}
+
+variable "execution_role_name" {
+  default     = null
+  description = "The arn of a user defined role that will be used as the execution role for the lambda function. "
+  type        = string
+}
+
+variable "reserved_concurrent_executions" {
+  default     = 0
+  description = "The number of concurrent executions that are reserved for this function, leave as default to use unreserved account concurrency"
+  type        = number
+
 variable "govcloud_deployment" {
   description = "Enable if you deploy the integration in govcloud"
   type        = bool